AI now works- Breathing now works for riding ground, riding air, and not riding.  Bite works.
